Fixup broken network code, broken settings & add reverse cmd line option
