+++ turntableClient/qtquickcontrols2.conf
@@ -7,9 +7,9 @@ Style=Material
 
 [Universal]
 Theme=Light
-;Accent=Steel
+Accent=Steel
 
 [Material]
 Theme=Light
-;Accent=BlueGrey
-;Primary=BlueGray
+Accent=LightBlue
+Primary=Blue
